[MOD] refactor problems and tips to a JSON-based repository and implement initial providers
- Added `assets/data/problems.json` containing multi-language pool problem procedures and general rules.
- Refactored `ProblemRepository` and `ProblemStepRepository` to use JSON data instead of SQLite, simplifying the domain layer.
- Updated `ProblemEntity` and `ProblemStepEntity` to use `String` IDs and simplified their structure.
- Implemented `Riverpod` providers for chemicals, inventory, measurements, problems, tips, and treatments.
- Updated `AppDatabase` schema (v5) to support decimal concentrations in chemicals and enhanced treatment tracking.
- Refactored `TreatmentEntity` and `TreatmentRepository` to include better filtering and relationship handling.
- Added placeholder screens for `History`, `Inventory`, `Calculators`, and `ProblemCategoryList`.
- Updated `LateralPoolMenu` to handle navigation state and dynamic pool data.

+++ lib/domain/entities/chemical_entity.dart
@@ -8,7 +8,7 @@ class ChemicalEntity {
   int? chemicalTypeId;
   ChemicalType? chemicalType;
   ChemicalFormat? format;
-  int? concentration;
+  double? concentration;
   Unit? unit;
 
   ChemicalEntity({
@@ -27,7 +27,7 @@ class ChemicalEntity {
     int? chemicalTypeId,
     ChemicalType? chemicalType,
     ChemicalFormat? format,
-    int? concentration,
+    double? concentration,
     Unit? unit,
   }) {
     return ChemicalEntity(
@@ -52,7 +52,7 @@ class ChemicalEntity {
       format: json['format'] != null
           ? ChemicalFormat.values.firstWhere((e) => e.name == json['format'])
           : null,
-      concentration: json['concentration'] as int?,
+      concentration: json['concentration'] as double?,
       unit: json['unit'] != null
           ? Unit.values.firstWhere((e) => e.name == json['unit'])
           : null,
